extract getSourceDataGroups out
